+++ R/delta.R
@@ -1,5 +1,5 @@
 extract_dmu <- function(x, outcome, estimand = "ATE") {
-  if (!inherits(x, "cea_estimate")) stop_not_cea_estimate()
+  if (!inherits(x, "cea_estimate")) stop_incorrect_class("cea_estimate")
   if (length(idx <- which(names(x$linear_pred) == outcome)) == 0) stop_unknown_outcome(outcome)
   tx <- attr(x, "tx")
   if (is.null(x$data[[tx]])) stop_unknown_treatment(tx)