+++ R/errors-extract.R
@@ -1,6 +1,6 @@
-stop_not_cea_estimate <- function() {
-  x <- "`x` must be an object of class 'cea_estimate'."
-  rlang::cnd_signal(rlang::error_cnd("cea_error_not_cea_estimate", message = x))
+stop_incorrect_class <- function(class) {
+  x <- paste0("`x` must be an object of class '", class, "'.")
+  rlang::cnd_signal(rlang::error_cnd("cea_error_incorrect_class", message = x))
 }
 
 stop_unknown_estimand <- function(estimand) {