+++ tests/testthat/setup.R
@@ -25,12 +25,20 @@ fit_lp <- estimate(
   data = moa2_ex
 )
 
+boot_est <- boot(fit, R = 9)
+boot_est_par <- boot(fit, R = 9, sim = "parametric")
+boot_est_fct <- boot(fit_fct, R = 9, sim = "parametric")
+boot_est_fct2 <- boot(fit_fct2, R = 9, sim = "parametric")
+
+### MI methods require `mice`
+
+skip_if_not_installed("mice", "3.0")
+
 moa2_mi <- rbind(moa2, moa2, moa2)
 moa2_mi$.imp <- as.character(rep(0:2, each = nrow(moa2)))
 moa2_mi <- mice::as.mids(moa2_mi)
 fit_mi <- estimate("QALYs", "Cost", "tx", c("age", "sex"), data = moa2_mi)
 
-boot_est <- boot(fit, R = 9)
-boot_est_par <- boot(fit, R = 9, sim = "parametric")
-boot_est_fct <- boot(fit_fct, R = 9, sim = "parametric")
-boot_est_fct2 <- boot(fit_fct2, R = 9, sim = "parametric")
+fit_pooled <- pool_cea(fit_mi)
+
+boot_pooled <- boot(fit_pooled, R = 9, sim = "parametric")