+++ tests/testthat/test-tidiers.R
@@ -26,3 +26,20 @@ test_that("tidy.cea_estimate works", {
   expect_equal(td_fct$statistic, td_fct$estimate / td_fct$std.error)
   expect_equal(td_fct$p.value, 2 * (1 - stats::pnorm(abs(td_fct$statistic))))
 })
+
+test_that("tidy.cea_pooled works", {
+  td <- tidy(fit_pooled)
+  expect_s3_class(td, "tbl_df")
+  expect_equal(dim(td), c(15, 7))
+  expect_equal(td$component, rep(c("regression", "correlation", "dispersion"), c(12, 1, 2)))
+  expect_equal(td$y.level, c(rep(c("QALYs", "Costs"), each = 6), NA, "QALYs", "Costs"))
+  expect_equal(
+    td$term,
+    c(rep(c("(Intercept)", "txExB", "txMT", "txMT + ExB", "age", "sex"), 2),
+      "rho12", "tau11", "tau21")
+  )
+  expect_equal(td$estimate, coef(fit_pooled))
+  expect_equal(td$std.error, unname(sqrt(diag(vcov(fit_pooled)))))
+  expect_equal(td$statistic, td$estimate / td$std.error)
+  expect_equal(td$p.value, 2 * (1 - stats::pnorm(abs(td$statistic))))
+})